Implement get sender total expense

+++ src/main/java/com/kinkl/paymentlistservice/dto/SenderTotalInfoDto.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class SenderTotalInfoDto {
-    private int senderId;
-    private int totallySent;
+    private Long senderId;
+    private Long totallySent;
 }
